+++ db/sqlc/expenses.sql.go
@@ -17,8 +17,8 @@ RETURNING id, group_id, payer_id, amount, description, date
 `
 
 type CreateExpenseParams struct {
-	GroupID     int32     `json:"group_id"`
-	PayerID     int32     `json:"payer_id"`
+	GroupID     int64     `json:"group_id"`
+	PayerID     int64     `json:"payer_id"`
 	Amount      string    `json:"amount"`
 	Description string    `json:"description"`
 	Date        time.Time `json:"date"`
@@ -82,8 +82,8 @@ RETURNING id, group_id, payer_id, amount, description, date
 
 type UpdateExpenseParams struct {
 	ID          int64     `json:"id"`
-	GroupID     int32     `json:"group_id"`
-	PayerID     int32     `json:"payer_id"`
+	GroupID     int64     `json:"group_id"`
+	PayerID     int64     `json:"payer_id"`
 	Amount      string    `json:"amount"`
 	Description string    `json:"description"`
 	Date        time.Time `json:"date"`
